+++ fabric/contrib/files.py
@@ -18,37 +18,104 @@ def exists(path, use_sudo=False, verbose=False):
     """
     Return True if given path exists on the current remote host.
 
-    If ``use_sudo`` is True, will use `sudo` instead of `run`.
-
-    `exists` will, by default, hide all output (including the run line, stdout,
-    stderr and any warning resulting from the file not existing) in order to
-    avoid cluttering output. You may specify ``verbose=True`` to change this
-    behavior.
+    For an explanation of the arguments ``use_sudo`` and ``verbose``
+    please refer to `test_file`.
     """
-    func = use_sudo and sudo or run
-    cmd = 'test -e %s' % _expand_path(path)
-    # If verbose, run normally
-    if verbose:
-        with settings(warn_only=True):
-            return not func(cmd).failed
-    # Otherwise, be quiet
-    with settings(hide('everything'), warn_only=True):
-        return not func(cmd).failed
+    return test_file(path, 'e', use_sudo, verbose)
+
+
+def is_readable(path, use_sudo=False, verbose=False):
+    """
+    Return True if given path exists and read permission is granted 
+    to the current user on the current remote host.
+
+    For an explanation of the arguments ``use_sudo`` and ``verbose``
+    please refer to `test_file`.
+    """
+    return test_file(path, 'r', use_sudo, verbose)
+
+
+def is_writable(path, use_sudo=False, verbose=False):
+    """
+    Return True if given path exists and write permission is granted
+    to the current user on the current remote host.
+
+    For an explanation of the arguments ``use_sudo`` and ``verbose``
+    please refer to `test_file`.
+    """
+    return test_file(path, 'w', use_sudo, verbose)
+
+
+def is_executable(path, use_sudo=False, verbose=False):
+    """
+    Return True if given path exists and execute (or search) permission is granted
+    to the current user on the current remote host.
+
+    For an explanation of the arguments ``use_sudo`` and ``verbose``
+    please refer to `test_file`.
+    """
+    return test_file(path, 'x', use_sudo, verbose)
+
+
+def is_dir(path, use_sudo=False, verbose=False):
+    """
+    Return True if given path exists and is a directory on the current remote
+    host.
+
+    For an explanation of the arguments ``use_sudo`` and ``verbose``
+    please refer to `test_file`.
+    """
+    return test_file(path, 'd', use_sudo, verbose)
+
+
+def is_file(path, use_sudo=False, verbose=False):
+    """
+    Return True if given path exists and is a regular file on the current remote
+    host.
+
+    For an explanation of the arguments ``use_sudo`` and ``verbose``
+    please refer to `test_file`.
+    """
+    return test_file(path, 'f', use_sudo, verbose)
 
 
 def is_link(path, use_sudo=False, verbose=False):
     """
     Return True if the given path is a symlink on the current remote host.
 
-    If ``use_sudo`` is True, will use `.sudo` instead of `.run`.
-
-    `.is_link` will, by default, hide all output. Give ``verbose=True`` to change this.
+    For an explanation of the arguments ``use_sudo`` and ``verbose``
+    please refer to `test_file`.
     """
+    return test_file(path, 'L', use_sudo, verbose)
+
+
+def test_file(path, op, use_sudo=False, verbose=False):
+    """
+    Task wrapping the `test` Unix command, specifically used for testing files.
+    Return True if the test against the file passes on the current remote host.
+
+    ``op`` is a unary operator for testing files, accepted by the `test` command.
+    The allowed values are `b`, `c`, `d`, `e`, `f`, `g`, `G`, `h`, `k`, `L`, `O`,
+    `p`, `r`, `s`, `S`, `u`, `w`, `x`.
+    Refer to `man test` for a detailed explanation.
+
+    If ``use_sudo`` is True, will use `sudo` instead of `run`.
+
+    `test_file` will, by default, hide all output (including the run line, stdout,
+    stderr and any warning resulting from the file not passing the test)
+    in order to avoid cluttering output. You may specify ``verbose=True``
+    to change this behavior.
+    """
+    # check that `op` is a valid unary operator for the `test` command
+    op_pattern = r'^[bcdefgGhkLOprsSuwx]$'
+    if not re.search(op_pattern, op):
+        raise ValueError("'%s' is not a valid operator for the 'test' Unix command" % op +
+                        " (please refer to `man test`)")
     func = sudo if use_sudo else run
-    cmd = 'test -L "$(echo %s)"' % path
+    cmd = 'test -%s %s' % (op, _expand_path(path))
     args, kwargs = [], {'warn_only': True}
     if not verbose:
-        opts = [hide('everything')]
+        args += [hide('everything')]
     with settings(*args, **kwargs):
         return func(cmd).succeeded