+++ docs/changelog.rst
@@ -4,6 +4,13 @@
 Changelog
 =========
 
+* :feature:`1047` Added `contrib.files.test_file <.test_file>` wrapping the `test`
+  Unix command for testing files, and the more specific tasks
+  `contrib.files.is_readable <.is_readable>`,
+  `contrib.files.is_writable <.is_writable>`,
+  `contrib.files.is_executable <.is_executable>`,
+  `contrib.files.is_dir <.is_dir>`, `contrib.files.is_file <.is_file>`.
+  Thanks to Alberto Scotto aka `@alb-i986` for the patch.
 * :release:`1.8.1 <2013-12-24>`
 * :release:`1.7.1 <2013-12-24>`
 * :release:`1.6.4 <2013-12-24>` 956, 957